requirements.txt
opencv-python
temp
streamlit
mediapipe
numpy
Pillow
time
os
pyaudio
SpeechRecognition
